Closes #1
Adds input type submit button so that the form is submitted on keypress of the enter key.  Adds a preventdefault to the scripting in the backend to prevent a page refresh and changes the selector for the login event from the ID to the class which is shared by both buttons.