Fixes #26 by rewriting how logs are tailed
The issue was caused because I was writing the tailed outcome to the log file and doing a comparison between that and the response from the server on the next call.  This caused it to never match and to display the full response.  This fixes that by creating a new tailed file which saves off the previous tailed results for display.  This code is very WET and in much need a refactor but it works for now.